+++ packages/wrapper/test/editorApp.test.ts
@@ -27,8 +27,8 @@ describe('Test EditorApp', () => {
 
     test('config defaults', () => {
         const wrapperConfig = createWrapperConfigExtendedApp();
-        const app = new EditorApp('config defaults', wrapperConfig.$type, wrapperConfig.editorAppConfig);
-        expect(app.getConfig().codeResources?.modified?.text).toEqual('');
+        const app = new EditorApp(wrapperConfig.$type, wrapperConfig.editorAppConfig);
+        expect(app.getConfig().codeResources?.modified?.text).toEqual('console.log("Hello World!");');
         expect(app.getConfig().codeResources?.original).toBeUndefined();
         expect(app.getConfig().useDiffEditor ?? false).toBeFalsy();
         expect(app.getConfig().readOnly).toBeFalsy();