+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this npm module are documented in this file.
 
+## [9.5.0-next.0] - 2025-03-13
+
+- Updated all `@codingame/monaco-vscode` packages to `15.0.2`.
+
 ## [9.4.1] - 2025-03-05
 
 - Updated all `@codingame/monaco-vscode` packages to `14.0.6`.